requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
opencv-python==4.10.0.84
rembg==2.0.57
ffmpeg-python==0.2.0
Pillow==10.4.0
celery==5.4.0
redis==5.0.8
aiofiles==24.1.0
python-dotenv==1.0.1
